style: format code and simplify code

+++ sort/inc/insert_sort.h
@@ -1,6 +1,5 @@
 #ifndef INSERT_SORT_H
 #define INSERT_SORT_H
 
-#include <stdlib.h>
-void insert_sort(int *array, size_t size);
+void insert_sort(int *array, int size);
 #endif
